\name{vanilla-package}
\alias{vanilla-package}
\alias{vanilla}
\docType{package}
\title{Tests the package loading implementation}
\description{Tests the package loading implementation}
\details{
\tabular{ll}{
Package: \tab vanilla\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-21\cr
License: \tab GPL-2\cr
}

}
\author{
The FastR Team

Maintainer: fastr@yahoogroups.com
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{

}
